refactor: move getCards method to moviesList component
